Add files via upload
This is a Python program for implementing ADT-GWR model. To utilize this program, you only need to update the lines introducing the data path. The program can read *.csv datasets with this structure: the X and Y in the first two columns, the dependent variable in the third column, and the independent variables in the next columns.
Four standard datasets have been incorporated to test the program. The data originates from https://github.com/Lessani252/SGWR.